+++ src/app/User/MakeReservation/page.tsx
@@ -1,39 +1,65 @@
-'use client'
-import React, { useState } from 'react'
-import Link from 'next/link';
-import { useRouter } from 'next/navigation';
+'use client';
+import React, { useState } from 'react';
+import axios from 'axios';
+import { useRouter } from 'next/navigation'; // Import useRouter for navigation
 
-export default function MakeReservation() {
+// Lambda function URL
+const LAMBDA_URL = 'https://42y3io3qm4.execute-api.us-east-1.amazonaws.com/Initial/createUser';
+
+export default function Authorization() {
   const [formData, setFormData] = useState({ name: '', email: '' });
-  const [error, setError] = useState('');
-  const router = useRouter(); // Next.js router for navigation
+  const router = useRouter(); // Initialize the router
 
   const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {
     const { name, value } = e.target;
     setFormData({ ...formData, [name]: value });
   };
 
-  const handleContinue = (e: React.FormEvent) => {
+  const handleSubmit = async (e: React.FormEvent) => {
     e.preventDefault();
-    if (!formData.name || !formData.email) {
-      setError('Both fields are required.');
+
+    const { name, email } = formData;
+
+    if (!name || !email) {
+      alert('Please provide a valid name and email');
       return;
     }
-    setError('');
-    console.log('Reservation details:', formData);
-    router.push('/ListActiveRestaurant'); // Redirect to ListActiveRestaurant page
+
+    try {
+      // Send POST request to Lambda function using Axios
+      const response = await axios.post(LAMBDA_URL, { name, email });
+
+      console.log('API Response:', response);
+
+      // Parse the body from the Lambda response
+      if (response.status === 200 && response.data) {
+        const parsedBody = JSON.parse(response.data.body); // Parse the body string
+        const { idUser } = parsedBody.data; // Extract idUser from the parsed body
+
+        // Store idUser in session storage
+        sessionStorage.setItem('idUser', idUser);
+        console.log('Generated idUser:', idUser);
+
+        // Redirect to chooseDate page
+        router.push('/User/chooseDate');
+      } else {
+        console.error('Unexpected API response:', response.data);
+        alert('Error: Unexpected API response structure.');
+      }
+    } catch (error) {
+      console.error('Error making request:', error);
+      alert('An unexpected error occurred.');
+    }
   };
 
   return (
     <main className="flex min-h-screen flex-col items-center justify-center p-6 bg-gray-100">
-      {/* Reservation Form */}
       <div className="w-full max-w-md bg-white shadow-lg rounded-lg p-8">
-        <h1 className="text-2xl font-bold mb-6 text-center text-gray-800">Make a Reservation</h1>
-        <form onSubmit={handleContinue} className="space-y-4">
-          {/* Name Input */}
-          <div>
-            <label htmlFor="name" className="block text-sm font-medium text-gray-700 mb-1">
-              Name
+        <h1 className="text-2xl font-bold mb-6 text-center text-black">Reservation Information</h1>
+        <form onSubmit={handleSubmit}>
+          <div className="mb-4">
+            <label className="block text-sm font-medium text-gray-700 mb-1" htmlFor="name">
+              First and Last Name
             </label>
             <input
               id="name"
@@ -42,14 +68,11 @@ export default function MakeReservation() {
               value={formData.name}
               onChange={handleInputChange}
               required
-              className="w-full px-3 py-2 border rounded-md shadow-sm focus:ring-blue-500 focus:border-blue-500"
-              placeholder="Enter your name"
+              className="w-full px-3 py-2 border rounded-md text-gray-700 shadow-sm focus:ring-blue-500 focus:border-blue-500"
             />
           </div>
-
-          {/* Email Input */}
-          <div>
-            <label htmlFor="email" className="block text-sm font-medium text-gray-700 mb-1">
+          <div className="mb-4">
+            <label className="block text-sm font-medium text-gray-700 mb-1" htmlFor="email">
               Email
             </label>
             <input
@@ -59,36 +82,16 @@ export default function MakeReservation() {
               value={formData.email}
               onChange={handleInputChange}
               required
-              className="w-full px-3 py-2 border rounded-md shadow-sm focus:ring-blue-500 focus:border-blue-500"
-              placeholder="Enter your email"
+              className="w-full px-3 py-2 border rounded-md text-gray-700 shadow-sm focus:ring-blue-500 focus:border-blue-500"
             />
           </div>
-
-          {/* Error Message */}
-          {error && <p className="text-red-500 text-sm">{error}</p>}
-
-          {/* Continue Button */}
-          <Link
-            href="/User/ListActiveRestaurant"
-            className="w-full block"
+          <button
+            type="submit"
+            className="w-full bg-blue-500 text-white py-2 px-4 rounded-md hover:bg-blue-600 transition duration-200"
           >
-            <button
-              type="submit"
-              className="w-full bg-green-500 text-white py-2 px-4 rounded-md hover:bg-green-600 transition duration-200"
-            >
-              Continue
-            </button>
-          </Link>
+            Confirm
+          </button>
         </form>
-
-        {/* Back Link */}
-        <div className="mt-4 text-center">
-          <Link href="/">
-            <button className="text-blue-500 hover:underline focus:outline-none">
-              Back to Home
-            </button>
-          </Link>
-        </div>
       </div>
     </main>
   );
